Universal File to Text Converter/requirements.txt
streamlit==1.26.0
Pillow==10.0.1
pytesseract==0.3.10
pdfplumber==0.9.0
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2